+++ tests/functions/test_fill_direction.py
@@ -88,28 +88,28 @@ def test_fill_column(df):
     """Fill down on a single column."""
     expected = df.copy()
     expected.loc[:, "pet_type"] = expected.loc[:, "pet_type"].ffill()
-    assert_frame_equal(df.fill_direction({"pet_type": "down"}), expected)
+    assert_frame_equal(df.fill_direction(**{"pet_type": "down"}), expected)
 
 
 def test_fill_column_up(df):
     """Fill up on a single column."""
     expected = df.copy()
     expected.loc[:, "pet_type"] = expected.loc[:, "pet_type"].bfill()
-    assert_frame_equal(df.fill_direction({"pet_type": "up"}), expected)
+    assert_frame_equal(df.fill_direction(**{"pet_type": "up"}), expected)
 
 
 def test_fill_column_updown(df):
     """Fill upwards, then downwards on a single column."""
     expected = df.copy()
     expected.loc[:, "pet_type"] = expected.loc[:, "pet_type"].bfill().ffill()
-    assert_frame_equal(df.fill_direction({"pet_type": "updown"}), expected)
+    assert_frame_equal(df.fill_direction(**{"pet_type": "updown"}), expected)
 
 
 def test_fill_column_down_up(df):
     """Fill downwards, then upwards on a single column."""
     expected = df.copy()
     expected.loc[:, "pet_type"] = expected.loc[:, "pet_type"].ffill().bfill()
-    assert_frame_equal(df.fill_direction({"pet_type": "downup"}), expected)
+    assert_frame_equal(df.fill_direction(**{"pet_type": "downup"}), expected)
 
 
 def test_fill_multiple_columns(df):
@@ -119,7 +119,7 @@ def test_fill_multiple_columns(df):
         :, ["pet_type", "owner"]
     ].ffill()
     assert_frame_equal(
-        df.fill_direction({"pet_type": "down", "owner": "down"}), expected
+        df.fill_direction(**{"pet_type": "down", "owner": "down"}), expected
     )
 
 
@@ -129,28 +129,29 @@ def test_fill_multiple_columns_multiple_directions(df):
     expected.loc[:, "pet_type"] = expected.loc[:, "pet_type"].ffill()
     expected.loc[:, "owner"] = expected.loc[:, "owner"].bfill()
     assert_frame_equal(
-        df.fill_direction({"pet_type": "down", "owner": "up"}), expected
+        df.fill_direction(**{"pet_type": "down", "owner": "up"}), expected
     )
 
 
 def test_wrong_column_name(df):
     """Raise Value Error if wrong column name is provided."""
     with pytest.raises(ValueError):
-        df.fill_direction({"PetType": "down"})
+        df.fill_direction(**{"PetType": "down"})
 
 
 def test_wrong_direction(df):
     """Raise Value Error if wrong direction is provided."""
     with pytest.raises(ValueError):
-        df.fill_direction({"pet_type": "upanddawn"})
+        df.fill_direction(**{"pet_type": "upanddawn"})
 
 
+@pytest.mark.xfail(reason="limit is deprecated")
 def test_wrong_type_limit(df):
     """Raise TypeError if limit is wrong type."""
     with pytest.raises(TypeError):
-        df.fill_direction({"pet_type": "up"}, limit="one")
+        df.fill_direction(**{"pet_type": "up"}, limit="one")
 
 
 def test_empty_directions(df):
     """Return dataframe if `directions` is empty."""
-    assert_frame_equal(df.fill_direction({}), df)
+    assert_frame_equal(df.fill_direction(), df)